Expand index traversal tests, remove formIndex(_:offsetBy:)
The current index traversal tests didn't invoke every single index
method, so we need a few additional tests. Also run them all on the
`Deque.Indices` collection too.

`formIndex(_:offsetBy:)` turns out not to actually be declared in any
stdlib collection protocol, instead it's just an extension on
`Collection`. Remove our implementations of it in favor of just using
the standard extension. This way the same code is invoked when calling
it on the `Deque` directly as when calling it on a generic type bound by
`Collection`.